backend/requirements.txt
fastapi
uvicorn[standard]
numpy
torch
transformers
python-multipart
python-dotenv
pyjwt
requests

